Move register model into launcher
